% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_commits.R
\name{compare_branch}
\alias{compare_branch}
\title{Run-times across branches.}
\usage{
compare_branch(test_path, branch_name, num_commits)
}
\arguments{
\item{test_path}{File-path for the test file to be tested.}

\item{branch_name}{Branch name against whose commits the test file is to be
tested.}

\item{num_commits}{Number of commits on the branch against which the test
  file is to be tested.}
}
\description{
Given a test-file and branch, returns the run-time of the file over the given
number of commits on the branch and the latest commit on master.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
compare_branch(test_path = t_path, branch_name = "helper", num_commits = 10)
}
\seealso{
\code{\link[git2r]{commits}}
}

