% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/branch_metrics.R
\name{compare_brancht}
\alias{compare_brancht}
\title{Run-times across branches.}
\usage{
compare_brancht(test_path, branch1, branch2 = "master")
}
\arguments{
\item{test_path}{File-path for the test file to be tested.}

\item{branch_1}{First Branch against whose commits the test file is to be
tested.}

\item{branch_2}{Second Branch against whose commits the test file is to be
  tested.}
}
\description{
Given a test-file and two branches, returns the run-times of the file over
the two branches against the first commit till the latest commit common to
both of them.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
compare_brancht(test_path = t_path, branch1 = "helper", branch2 = "master")
}

