% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travisPR_metrics.R
\name{compare_dir}
\alias{compare_dir}
\title{Compare performance across directories/repositories.}
\usage{
compare_dir(dir1, dir2, test_path, metric = "time", PR = F)
}
\arguments{
\item{dir1}{Path to the first directory/repository.}

\item{dir2}{Path to the second directory/repository.}

\item{test_path}{File-path, relative to the directories, for the test file to
be tested.}

\item{metric}{The metric (runtime or memory) for which the file is to be 
tested.}

\item{PR}{When set to True, it performs the analysis in accordance with the setup
on Travis-CI environment for testing Pull Requests.}
}
\description{
Given a test-file, two directories and the required metric, returns the 
metric details of the file against the first commit till the latest common 
commit for dir1, and against the latest commit for dir2. It also returns
information regarding the latest common commit for the two directories.
}
\section{Value}{
 compare_brancht returns an object of class "list" containing
  two members.

The first member is a data-frame and consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{sha}
\code{date_time}
\code{directory}

The second member is a data-frame and consists of the following columns:
\code{common_datetime}
\code{common_message}
\code{cnum_b1}
\code{cnum_b2}
}

\section{Warning}{

  Function assumes the current directory to be the parent directory of both 
  the repositories being tested. That means both the repositories should be
  inside the same directory.
}
\examples{

\dontrun{
# Set the current directory to the parent directory of the concerned repositories.
setwd("./Path/to/parent/directory")

# Set the directory paths
d_path1 <- "Path/to/first/directory"
d_path2 <- "Path/to/second/directory"

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
compare_dir(d_path1, d_path2, t_path, metric = "time")
}

}

