% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/branch_metrics.R
\name{compare_dirt}
\alias{compare_dirt}
\title{Run-time details across directories/repositories.}
\usage{
compare_dirt(dir1, test_path1, branch1 = "master", dir2, test_path2,
  branch2 = "master")
}
\arguments{
\item{dir1}{Path to the first directory/repository.}

\item{test_path1}{File-path, relative to the first directory, for the test
file to be tested.}

\item{dir2}{Path to the second directory/repository.}

\item{test_path2}{File-path, relative to the second directory, for the test
file to be tested.}

\item{branch_1}{Branch in the first repository against whose commits the test
file is to be tested.}

\item{branch_2}{Branch in the second directory against whose commits the test
  file is to be tested.}
}
\description{
Given a test-file and two branches, returns the run-time details of the file
against the first commit till the latest common commit in branch1, and
against the latest commit in branch2.
}
\section{Value}{

compare_brancht returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{date_time}
\code{branch}
\code{directory}
}

\section{Warning}{

  Function assumes the current directory to be the parent directory of both
  the repositories being tested. That means both the repositories should be
  inside the same directory.
}
\examples{
\dontrun{
# Set the current directory to the parent directory of the concerned repositories.
setwd("./Path/to/parent/directory")

# Set the directory paths
d_path1 <- "Path/to/first/directory"
d_path2 <- "Path/to/second/directory"

# Set the file-paths
t_path1 <- "First/path/to/file"
t_path2 <- "Second/path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
compare_dirt(d_path1, t_path1, branch1 = "master",
             d_path2, t_path2, branch2 = "patch")
}
}

