% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_pr_comment}
\alias{create_pr_comment}
\title{Creates a Pull Request Comment file}
\usage{
create_pr_comment(test_results, test_function, target_dir, folder_name)
}
\arguments{
\item{test_results}{A Dataset.}

\item{test_function}{A String.}

\item{target_dir}{A String.}

\item{folder_name}{A String.}
}
\description{
Creates a Pull Request Comment file
}
