% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_comment_text}
\alias{get_comment_text}
\title{Modifying the PR Comment}
\usage{
get_comment_text(part = c("footer", "header"), default, env = parent.frame())
}
\description{
The files `rperform/header.R` and `rperform/footer.R` allow you to modify
the PR comment. You can use github markdown syntax to format the text.
}
