% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_help.R
\name{get_datetime}
\alias{get_datetime}
\title{DateTime value of a git commit object.}
\usage{
get_datetime(commit_val)
}
\arguments{
\item{commit_val}{git commit object, as returned by git2r::commits()}
}
\description{
\code{get_sha(commit_val = )} returns the date-time value for the git commit
object provided as the parameter.
}
\seealso{
\code{\link[git2r]{commits}}
}

