% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_metrics.R
\name{get_mem}
\alias{get_mem}
\title{Test-file's memory statistics.}
\usage{
get_mem(test_path, commit_num = 1)
}
\arguments{
\item{test_path}{File-path for the test file which is to be checked.}

\item{commit_num}{commit number in the git log for the current git repository
against which the memory stats are to be checked, with the commit number
for the most recent commit being 1.}
}
\description{
Given a test-file's path, checks its memory metrics against the commit 
specified by the commit number passed as a parameter. Memory metrics returned
are the memory leaked and maximum memory utilized during its execution. A
commit number,n, would correspond to the nth commit in the commit log of the
current git repository.
}
\section{Value}{

get_mem returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{sha}
\code{date_time}
}

\section{Warning}{

  Function assumes the current directory to be the root directory of the
  package being tested.
}
\examples{

\dontrun{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the memory stats
library(Rperform)
get_mem(t_path, 3)
}

}

