% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_help.R
\name{get_sha}
\alias{get_sha}
\title{SHA1 value of a git commit object.}
\usage{
get_sha(commit_val)
}
\arguments{
\item{commit_val}{git commit object, as returned by git2r::commits()}
}
\description{
\code{get_sha(commit_val = )} returns the SHA1 value for the git commit object provided
as the parameter.
}
\seealso{
\code{\link[git2r]{commits}}
}

