% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_commits.R
\name{get_times}
\alias{get_times}
\title{Run-time across versions.}
\usage{
get_times(test_path, num_commits = 20, mode = 1)
}
\arguments{
\item{test_path}{File-path of the test-file which is to be used for run-time
comparisons.}

\item{num_commits}{Number of commits (versions) against which the file is to
  be tested, with default being 20.}
}
\description{
Given a test-file path, checks its run-time against the specified number of commits
in the current git repository and returns a data-frame comprised of the test name,
status of test run, time (if successful) and SHA1 value corresponding to the commit
the value is for.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details against 10 commits
library(Rperform)
time_commit(test_path = t_path, n_commits = 10)
}
\seealso{
\code{\link[git2r]{commits}}
}

