% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo_metrics.R
\name{list_commits}
\alias{list_commits}
\title{Commits' details.}
\usage{
list_commits(path = "./", num_commits = 20)
}
\arguments{
\item{path}{File-path to the git repository whose commits are to be summarized.}

\item{num_commits}{Number of commits to be summarized. The default is 20.}
}
\description{
Given a repository path and number of commits (n), returns a data frame containing
the date, SHA1 values and commit messages of the last n commits in the repo.
}
\examples{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Obtained details of the last 10 commits in the repository
list_commits(num_commits = 10)

## Example-2

# Obtained the details of the last 20 (default value) commits in the repository
# specified by path.
list_commits(path)
}

