% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_commits.R
\name{mem_compare}
\alias{mem_compare}
\title{Test-file's memory statistics for multiple commits.}
\usage{
mem_compare(test_path, num_commits = 5)
}
\arguments{
\item{test_path}{File-path for the test file which is to be checked.}

\item{num_commits}{number of commits against all of which the memory stats
  are to be checked starting from the most recent one.}
}
\description{
Given a test-file's path, checks its memory metrics against the number of
commits specified by the parameter num_commits. Memory metrics returned are
the memory leaked and maximum meory swapped during its execution.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details
library(Rperform)
mem_commit(t_path, 5)
}
\seealso{
\code{\link[git2r]{commits}}
}

