% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_bmemory}
\alias{plot_bmemory}
\title{Plot memory metrics across branches.}
\usage{
plot_bmemory(test_path, branch1, branch2 = "master")
}
\arguments{
\item{test_path}{File-path for the test file to be tested.}

\item{branch1}{Branch against whose commits the test file is to be 
tested.}

\item{branch2}{Branch into which branch1 is supposedly to be merged.}
}
\description{
Given a test-file and two branches, plots the memory metrics of the file 
against the first commit till the latest common commit in branch1, and 
against the latest commit in branch2. The memory metrics plotted are the 
memory leaked and maximum swapped memory during execution. The vertical line 
divides the commits from the two branches with the ones from branch1 on the 
left side.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{

\dontrun{
# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
plot_bmemory(test_path = t_path, branch1 = "helper", branch2 = "master")
}

}

