% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_html}
\alias{plot_html}
\title{Plot time and memory metrics of all test files on a webpage}
\usage{
plot_html(test_directory = "tests/testthat", output_name = "index")
}
\arguments{
\item{test_directory}{Path of the directory containing the test files.}

\item{out_file}{Name of the output webpage.}
}
\description{
It plots the time and memory metrics for all the tests present in the specified
directory of the current git repository on a webpage.
}
\section{WARNING}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Set to the git repository in consideration.
setwd("path/to/repo")

# Load the package and run the function
library(Rperform)
plot_html(test_directory = "path/to/tests", out_file = "output.Rmd")
}

