% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_mem}
\alias{plot_mem}
\title{Plot memory statistics across versions.}
\usage{
plot_mem(test_path, num_commits = 5, save_data = FALSE)
}
\arguments{
\item{test_path}{File-path of the test-file which is to be used for run-time
comparisons.}

\item{num_commits}{Number of commits (versions) against which the file is to
be tested, with default being 5.}

\item{save_data}{If set to TRUE, the data frame containing the memory metrics
  is stored.}
}
\description{
Given a test-file path, plot the memory usage statistics of entire file
against the commit message summaries of the specified number of commits in
the current git repository. The statistics plotted are the memory leaked and
memory swapped during the run of a test file.
}
\section{WARNING}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the memory usage details against 10 commits
library(Rperform)
plot_mem(test_path = t_path, n_commits = 10)
}

