% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travisPR_metrics.R
\name{plot_PR_webpage}
\alias{plot_PR_webpage}
\title{Generate a webpage containing a visualization detailing PR's impact on
performance without haveing to merge.}
\usage{
plot_PR_webpage(test_path, metric = "time")
}
\arguments{
\item{test_path}{File-path, relative to the git repo, for the test file to
be tested.}

\item{metric}{The metric (runtime or memory) for which the file is to be 
tested.}
}
\description{
The function must be called from a directory containing only a single git
repository checked out to the branch which is meant to be tested against the
master branch of the repository's remote repo. This function is designed keeping
in mind the PR testing methodlogy of Travis-CI.
Given a test-file path and the required metric, it creates a webpage
visualizing the metric details of the file against the first commit till the
latest common commit for the git repo, and against the latest commit for
master branch of the repo's remote.
}
\section{Value}{
 None
}

\section{Warning}{

  Function assumes the current directory to be the parent directory of the 
  the repository being tested.
}
\examples{

\dontrun{
# Set the current directory to the parent directory of the concerned repository.
setwd("./Path/to/parent/directory")

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
plot_PR_webpage(t_path, metric = "time")
}

}

