% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_time}
\alias{plot_time}
\title{Plot run-time across versions.}
\usage{
plot_time(test_path, num_commits = 5, save_data = FALSE)
}
\arguments{
\item{test_path}{File-path of the test-file which is to be used for run-time
comparisons.}

\item{num_commits}{Number of commits (versions) against which the file is to
be tested, with default being 5.}

\item{save_data}{If set to TRUE, the data frame containing the time metrics
  information is stored.}
}
\description{
Given a test-file path, plot the run-time of entire file and individual
testthat blocks against the commit message summaries of the specified number
of commits in the current git repository. If the parameter save_data is set
to true, it also stores the corresponding data frame in an RData file in a
folder 'Rperform_Data' in the current directory.
}
\section{WARNING}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details against 10 commits
library(Rperform)
plot_time(test_path = t_path, n_commits = 10, save_data = F)
}

