% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_dir}
\alias{prepare_dir}
\title{Creates a new directory if it doesn't exist.}
\usage{
prepare_dir(new_dir)
}
\arguments{
\item{new_dir}{The directory to create.}
}
\description{
Creates a new directory if it doesn't exist.
}
