\name{Rperform-package}
\alias{Rperform-package}
\alias{Rperform}
\docType{package}
\title{
\packageTitle{Rperform}
}
\description{
\packageDescription{Rperform}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rperform}
\packageIndices{Rperform}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rperform}

Maintainer: \packageMaintainer{Rperform}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[https://github.com/ropensci/git2r]{git2r}} ~~
}
\examples{
}
