% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_rperform.R
\name{run_script}
\alias{run_script}
\title{Runs the Rperform script.}
\usage{
run_script(path = "rperform/script.R")
}
\arguments{
\item{path}{A String.}
}
\description{
Runs the Rperform script.
}
\examples{
\dontrun{
run_script("inst/script.R")
} 
}
