% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_metrics.R
\name{time_branch}
\alias{time_branch}
\title{Run-times of a file on the given branch.}
\usage{
time_branch(test_path, branch = "master", num_commits = 5)
}
\arguments{
\item{test_path}{File-path for the test file to be tested.}

\item{branch}{Branch against whose commits the test file is to be 
tested (with master being the default).}

\item{num_commits}{Number of commits on the branch against which the test
file is to be tested.}
}
\description{
Given a test-file and branch, returns the run-time details of the file over
the given number of commits on the branch.
}
\section{Value}{

time_branch returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{date_time}
\code{branch}
}

\section{Warning}{

  Function assumes the current directory to be the root directory of the package being tested.
}
\examples{

\dontrun{
# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
time_branch(test_path = t_path, branch_name = "helper", num_commits = 10)
}
}

