% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_commits.R
\name{time_commit1}
\alias{time_commit1}
\title{Test file's run-time (mode-1).}
\usage{
time_commit1(test_path, test_commit)
}
\arguments{
\item{test_path}{File-path for the test file whose run-time is to be checked.}

\item{test_commit}{git2r commit \code{object} corresponding to the commit corresponding to
   which the run-time is to be checked.}
}
\description{
Given a test-file's path, checks its run-time against the commit specified by the
commit \code{object} passed as a parameter.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1
# Obtain the commit object
commit_list <- git2r::commits()
t_commit <- commit_list[[1]]

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details
library(Rperform)
time_commit(t_path, t_commit)
}
\seealso{
\code{\link[git2r]{commits}}
}

