% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_rbuildignore}
\alias{append_rbuildignore}
\title{Appends dir to Rbuildignore}
\usage{
append_rbuildignore(dir)
}
\arguments{
\item{dir}{A String.}
}
\value{
A cli object
}
\description{
Appends dir to Rbuildignore
}
\examples{
append_rbuildignore("Rperform")
}
