% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_metrics.R
\name{compare_branchm}
\alias{compare_branchm}
\title{Memory metrics across branches.}
\usage{
compare_branchm(test_path, branch1, branch2 = "master")
}
\arguments{
\item{test_path}{File-path for the test file to be tested.}

\item{branch1}{Branch against whose commits the test file is to be 
tested.}

\item{branch2}{Branch into which branch1 is supposedly to be merged.}
}
\description{
Given a test-file and two branches, returns the memory metrics of the file 
against the first commit till the latest common commit in branch1, and 
against the latest commit in branch2. Memory metrics returned are the memory 
leaked and maximum meory swapped during its execution.
}
\section{Value}{

compare_branchm returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{date_time}
\code{time_branch}
}

\section{Warning}{

  Function assumes the current directory to be the root directory of the
  package being tested.
}
\examples{

\dontrun{
# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Set the file-path
t_path <- "Path/to/file"

# Load the library and pass the parameters to the function
library(Rperform)
compare_branchm(test_path = t_path, branch1 = "helper", branch2 = "master")
}

}

