% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_rperform.R
\name{copy_if_not_exists}
\alias{copy_if_not_exists}
\title{Copies a file if it doesn't exist.}
\usage{
copy_if_not_exists(path, new_path, overwrite = FALSE)
}
\description{
Copies a file if it doesn't exist.
}
