% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_help.R
\name{get_branch}
\alias{get_branch}
\title{Current branch name of a git repository.}
\usage{
get_branch(dir_path = "./")
}
\description{
\code{get_branch} returns the branch name for the git repository passed in as
parameter (default being the current repository).
}
\seealso{
\code{\link[git2r]{head}}

\code{\link[git2r]{repository}}
}

