% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo_metrics.R
\name{mem_commit}
\alias{mem_commit}
\title{Test-file's memory statistics.}
\usage{
mem_commit(test_path, test_commit)
}
\arguments{
\item{test_path}{File-path for the test file which is to be checked.}

\item{test_commit}{git2r commit \code{object} corresponding to which the
  memory stats are to be checked.}
}
\description{
Given a test-file's path, checks its memory metrics against the commit
specified by the commit \code{object} passed as a parameter. Memory
metrics returned are the memory leaked and maximum memory swapped during
its execution.
}
\section{Value}{

mem_commit returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{date_time}
}

\section{Warning}{

  Function assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Obtain the commit object
commit_list <- git2r::commits()
t_commit <- commit_list[[1]]

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the memory stats
library(Rperform)
mem_commit(t_path, t_commit)
}
\seealso{
\code{\link[git2r]{commits}}
}

