% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_metrics.R
\name{mem_compare}
\alias{mem_compare}
\title{Test-file's memory statistics for multiple commits.}
\usage{
mem_compare(test_path, num_commits = 10)
}
\arguments{
\item{test_path}{File-path for the test file which is to be checked.}

\item{num_commits}{number of commits against all of which the memory stats
are to be checked starting from the most recent one.}
}
\description{
Given a test-file's path, checks its memory metrics against the number of 
commits specified by the parameter num_commits with default being 10. Memory
metrics returned are the memory leaked and maximum meory utilized during its
execution.
}
\section{Value}{

time_commit returns an object of class "data.frame".
The data-frame consists of the following columns:
\code{test_name}
\code{metric_name}
\code{status}
\code{metric_val}
\code{message}
\code{sha}
\code{date_time}
}

\section{Warning}{

  Function assumes the current directory to be the root directory of the
  package being tested.
}
\examples{

\dontrun{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details
library(Rperform)
mem_compare(t_path, 10)
}

}

