% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_directory}
\alias{plot_directory}
\title{Plot memory and time statistics across versions for all files in a given directory.}
\usage{
plot_directory(test_dir, num_commits = 5)
}
\arguments{
\item{test_dir}{Directory containing the test-files which are to be used.}

\item{num_commits}{Number of commits (versions) against which the files are to
  be tested, with default being 5.}
}
\description{
Given a directory path, plot the memory and time usage statistics of all files
in the directory against the commit message summaries of the specified number
of commits in the current git repository.
}
\section{WARNING}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Set the current directory to the git repository concerned.
setwd("./Path/to/repository")

# Specify the directory containing the test files
d_path <- "Path/to/directory"

# Pass the parameters and obtain the memory usage details against 10 commits
library(Rperform)
plot_directory(test_dir = d_path, n_commits = 10)
}

