% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_time}
\alias{plot_time}
\title{Plot run-time across versions.}
\usage{
plot_time(test_path, num_commits)
}
\arguments{
\item{test_path}{File-path of the test-file which is to be used for run-time
comparisons.}

\item{num_commits}{Number of commits (versions) against which the file is to
  be tested, with default being 20.}
}
\description{
Given a test-file path, plot the run-time of individual testthat blocks
against the date-time values of the specified number of commits (with the
commits' date-time values on the y-axis) in the current git repository.
}
\section{Warning}{

  Library assumes the current directory to be the root directory of the
  package being tested.
}
\examples{
## Example-1

# Specify the test-file path
t_path <- "Path/to/file"

# Pass the parameters and obtain the run-time details against 10 commits
library(Rperform)
plot_time(test_path = t_path, n_commits = 10)
}
\seealso{
\code{\link[git2r]{commits}}

\code{\link[testthat]}
}

