\name{Rperform-package}
\alias{Rperform-package}
\alias{Rperform}
\docType{package}
\title{
\packageTitle{Rperform}
}
\description{
\packageDescription{Rperform}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rperform}
\packageIndices{Rperform}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rperform}

Maintainer: \packageMaintainer{Rperform}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
